Add files via upload
I added the basic particle system code unto Tile.lua, referenced the file in Dependencies.lua but I could not find a successful way to implement it within Playstate.lua or Board.lua as when I try to,  the game runs as expected but no particles appear when a match is made. I believe the reason to it not working is that I am trying to create an action unto a object that disappears before the action can take place, while with  the Breakout project, the brick would still exist after the ball touched it and could initiate the particles dispersing. With Match 3, the tiles disappear once the match happens and cant make the particles appear. This is an interesting dilemma and or theory that I would like your opinion and or advice on. I hope to one day implement this correctly and contribute to this project.